Add RandomNumber Draw
Adds code to handle creation, getting and tossing a draw of type random
number. The code is not supposed to be generic but just a basic PoC. A
following commit will refactor it to prepare for future draws.

Migrations code has not be included as the db is not finalized.